Move ChatRoadmap to pages directory
